* Use "-fPIC" instead of "-mrelocatable" to prevent problems with   recent tools
* Add checksum verification to 'imls' command

* Add bd_info fields needed for 4xx Linux I2C driver

* Patch by Martin Krause, 4 Nov. 2003:
  Fix error in cmd_vfd.c (TRAB board: "vfd /1" shows now only one Bitmap)

* Print used network interface when CONFIG_NET_MULTI is set

+++ tools/mkimage.c
@@ -291,7 +291,7 @@ NXTARG:		;
 			exit (EXIT_FAILURE);
 		}
 
-		if (sbuf.st_size < sizeof(image_header_t)) {
+		if ((unsigned)sbuf.st_size < sizeof(image_header_t)) {
 			fprintf (stderr,
 				"%s: Bad size: \"%s\" is no valid image\n",
 				cmdname, imagefile);
@@ -530,7 +530,7 @@ copy_file (int ifd, const char *datafile, int pad)
 		 * reserved for it.
 		 */
 
-		if (sbuf.st_size < sizeof(image_header_t)) {
+		if ((unsigned)sbuf.st_size < sizeof(image_header_t)) {
 			fprintf (stderr,
 				"%s: Bad size: \"%s\" is too small for XIP\n",
 				cmdname, datafile);
